+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/asset_publisher.ftl
@@ -1,7 +1,7 @@
 <#assign
 	assetClassNameIds = classNameDataFactory.assetClassNameIds
 	assetVocabularyModels = assetDataFactory.newAssetVocabularyModels(groupId)
-	pageCounts = dataFactory.getSequence(assetDataFactory.maxAssetPublisherPageCount)
+	pageCounts = counterDataFactory.getSequence(assetDataFactory.maxAssetPublisherPageCount)
 />
 
 <#list assetVocabularyModels as assetVocabularyModel>
