+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/blogs.ftl
@@ -22,7 +22,7 @@
 		_entry=blogsEntryModel
 	/>
 
-	<#assign mbRootMessageId = dataFactory.getCounterNext() />
+	<#assign mbRootMessageId = counterDataFactory.getCounterNext() />
 
 	<@insertMBDiscussion
 		_classNameId=blogsEntryClassNameId
@@ -30,7 +30,7 @@
 		_groupId=groupId
 		_maxCommentCount=blogDataFactory.maxBlogsEntryCommentCount
 		_mbRootMessageId=mbRootMessageId
-		_mbThreadId=dataFactory.getCounterNext()
+		_mbThreadId=counterDataFactory.getCounterNext()
 	/>
 
 	${dataFactory.toInsertSQL(dataFactory.newSubscriptionModel(blogsEntryModel, blogsEntryClassNameId))}
