+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/counters.ftl
@@ -1,4 +1,4 @@
-<#assign counterModels = dataFactory.newCounterModels() />
+<#assign counterModels = counterDataFactory.newCounterModels() />
 
 <#list counterModels as counterModel>
 	<#if '${counterModel.name}' == 'com.liferay.counter.kernel.model.Counter'>
